Added multithreading merger

+++ splitter-service/src/main/java/net/bondar/utils/SplitIterator.java
@@ -11,9 +11,9 @@ public class SplitIterator implements Iterable{
 
     private final static Logger log = Logger.getLogger(SplitIterator.class);
     private int partCounter = 1;
-    private int currentPosition = 0;
-    private int fileLength;
-    private int partLength;
+    private long currentPosition = 0;
+    private long fileLength;
+    private long partLength;
 
     public SplitIterator(int fileLength, int partLength) {
         this.fileLength = fileLength;
@@ -22,8 +22,8 @@ public class SplitIterator implements Iterable{
 
     public synchronized Task getNext() {
         if(currentPosition>fileLength) return new Task();
-        int start = currentPosition;
-        int end = currentPosition + partLength;
+        long start = currentPosition;
+        long end = currentPosition + partLength;
         if (end > fileLength) end = fileLength;
         String partName = "_part_" + String.format("%03d", partCounter++);
         Task task = new Task(partName, start, end, partCounter-1);
